Improve document symbols

+++ common/src/lib.rs
@@ -5,7 +5,7 @@ pub mod project_kind;
 
 #[derive(Debug, Default, Clone)]
 pub struct TaskProgress {
-    pub persentage: usize,
+    pub persentage: u32,
     pub error: bool,
     pub message: String,
 }
